Issue/489: wip
* Refactored sortable list. It is now much simpler, as most of its logic now resides in the composite sortable list, instead.
* Resolved a few cases of missing element IDs.
