+++ presentation/sheet/actor/part/abilities/actor-skills-viewmodel.mjs
@@ -8,10 +8,9 @@ import { ValidationUtil } from "../../../../../business/util/validation-utility.
 import { ExtenderUtil } from "../../../../../common/extender-util.mjs"
 import SpecificDocumentCreationStrategy from "../../../../../business/document/creation/specific-document-creation-strategy.mjs"
 import { ContextMenuItem } from "../../../../component/button-context-menu/button-context-menu-viewmodel.mjs"
-import CompositeSortableListViewModel from "../../../../component/composite-sortable-list/composite-sortable-list-viewmodel.mjs"
+import CompositeSortableListViewModel, { SortableListAddItemParams } from "../../../../component/composite-sortable-list/composite-sortable-list-viewmodel.mjs"
 import { SortingOption } from "../../../../component/sort-controls/sort-controls-viewmodel.mjs"
 import DocumentListItemOrderDataSource from "../../../../component/sortable-list/document-list-item-order-datasource.mjs"
-import { SortableListAddItemParams } from "../../../../component/sortable-list/sortable-list-viewmodel.mjs"
 import ViewModel from "../../../../view-model/view-model.mjs"
 import SkillListItemViewModel from "../../../item/skill/skill-list-item-viewmodel.mjs"
 import SyntheticRollStrategy from "../../npc/synthetic-roll-strategy.mjs"
