+++ presentation/sheet/actor/part/abilities/momentum/actor-momentum-viewmodel.mjs
@@ -2,10 +2,9 @@ import { ITEM_TYPES } from "../../../../../../business/document/item/item-types.
 import { StringUtil } from "../../../../../../business/util/string-utility.mjs";
 import { ValidationUtil } from "../../../../../../business/util/validation-utility.mjs";
 import SpecificDocumentCreationStrategy from "../../../../../../business/document/creation/specific-document-creation-strategy.mjs";
-import CompositeSortableListViewModel from "../../../../../component/composite-sortable-list/composite-sortable-list-viewmodel.mjs";
+import CompositeSortableListViewModel, { SortableListAddItemParams } from "../../../../../component/composite-sortable-list/composite-sortable-list-viewmodel.mjs";
 import { SortingOption } from "../../../../../component/sort-controls/sort-controls-viewmodel.mjs";
 import DocumentListItemOrderDataSource from "../../../../../component/sortable-list/document-list-item-order-datasource.mjs";
-import { SortableListAddItemParams } from "../../../../../component/sortable-list/sortable-list-viewmodel.mjs";
 import ViewModel from "../../../../../view-model/view-model.mjs";
 import MomentumActionListItemViewModel from "../../../../item/momentum-action/momentum-action-list-item-viewmodel.mjs";
