+++ presentation/sheet/actor/part/projects/actor-projects-viewmodel.mjs
@@ -3,9 +3,8 @@ import { StringUtil } from "../../../../../business/util/string-utility.mjs";
 import { ValidationUtil } from "../../../../../business/util/validation-utility.mjs";
 import { ExtenderUtil } from "../../../../../common/extender-util.mjs";
 import SpecificDocumentCreationStrategy from "../../../../../business/document/creation/specific-document-creation-strategy.mjs";
-import CompositeSortableListViewModel from "../../../../component/composite-sortable-list/composite-sortable-list-viewmodel.mjs";
+import CompositeSortableListViewModel, { SortableListAddItemParams } from "../../../../component/composite-sortable-list/composite-sortable-list-viewmodel.mjs";
 import DocumentListItemOrderDataSource from "../../../../component/sortable-list/document-list-item-order-datasource.mjs";
-import { SortableListAddItemParams } from "../../../../component/sortable-list/sortable-list-viewmodel.mjs";
 import ViewModel from "../../../../view-model/view-model.mjs";
 import ProjectListItemViewModel from "../../../item/project/project-list-item-viewmodel.mjs";
